refactor: Use a common execution function
